[ALPS03714648] wlan: Fix incompatible args in DBGLOG
[Detail]
This patch only fix coverities with incompatible args in DBGLOG,
which includes:
BORA00012431 for CID 70371
BORA00012792 for CID 4140463
BORA00012788 for CID 4140459
BORA00012768 for CID 4140443
BORA00012661 for CID 361862
BORA00012791 for CID 4140462

MTK-Commit-Id: ea218e844de2c62cda11b5b81c5806b1798a1bc3

CR-Id: ALPS03714648
Feature: Others
(cherry picked from commit 4c302b4e1fa81c35139dbc4b464350d010f03e5c)

+++ mgmt/saa_fsm.c
@@ -994,7 +994,7 @@ uint32_t saaFsmRunEventRxDeauth(IN struct ADAPTER *prAdapter, IN struct SW_RFB *
 					DBGLOG(RSN, INFO,
 					       "QM RX MGT: Deauth frame, P=%d Sec=%d CM=%d BC=%d fc=%02x\n",
 					       prAisSpecBssInfo->fgMgmtProtection,
-					       HAL_RX_STATUS_GET_SEC_MODE(prSwRfb->prRxStatus),
+					       (uint8_t) HAL_RX_STATUS_GET_SEC_MODE(prSwRfb->prRxStatus),
 					       HAL_RX_STATUS_IS_CIPHER_MISMATCH(prSwRfb->prRxStatus),
 					       IS_BMCAST_MAC_ADDR(prDeauthFrame->aucDestAddr),
 					       prDeauthFrame->u2FrameCtrl);
@@ -1178,7 +1178,7 @@ uint32_t saaFsmRunEventRxDisassoc(IN struct ADAPTER *prAdapter, IN struct SW_RFB
 					DBGLOG(RSN, INFO,
 					       "QM RX MGT: Disassoc frame, P=%d Sec=%d CM=%d BC=%d fc=%02x\n",
 					       prAisSpecBssInfo->fgMgmtProtection,
-					       HAL_RX_STATUS_GET_SEC_MODE(prSwRfb->prRxStatus),
+					       (uint8_t) HAL_RX_STATUS_GET_SEC_MODE(prSwRfb->prRxStatus),
 					       HAL_RX_STATUS_IS_CIPHER_MISMATCH(prSwRfb->prRxStatus),
 					       IS_BMCAST_MAC_ADDR(prDisassocFrame->aucDestAddr),
 					       prDisassocFrame->u2FrameCtrl);
